Move fragments to ui
