+++ app/src/main/java/com/amaze/filemanager/activities/superclasses/ThemedActivity.java
@@ -21,10 +21,10 @@
 package com.amaze.filemanager.activities.superclasses;
 
 import com.amaze.filemanager.R;
-import com.amaze.filemanager.fragments.preference_fragments.PreferencesConstants;
 import com.amaze.filemanager.ui.colors.ColorPreferenceHelper;
 import com.amaze.filemanager.ui.colors.UserColorPreferences;
 import com.amaze.filemanager.ui.dialogs.ColorPickerDialog;
+import com.amaze.filemanager.ui.fragments.preference_fragments.PreferencesConstants;
 import com.amaze.filemanager.utils.theme.AppTheme;
 
 import android.os.Build;
