+++ app/src/main/java/com/amaze/filemanager/asynchronous/services/AbstractProgressiveService.java
@@ -26,7 +26,7 @@ import com.amaze.filemanager.R;
 import com.amaze.filemanager.activities.MainActivity;
 import com.amaze.filemanager.asynchronous.management.ServiceWatcherUtil;
 import com.amaze.filemanager.filesystem.HybridFile;
-import com.amaze.filemanager.fragments.ProcessViewerFragment;
+import com.amaze.filemanager.ui.fragments.ProcessViewerFragment;
 import com.amaze.filemanager.ui.notifications.NotificationConstants;
 import com.amaze.filemanager.utils.DatapointParcelable;
 import com.amaze.filemanager.utils.ProgressHandler;
