+++ app/src/main/java/com/amaze/filemanager/ui/fragments/AppsListFragment.java
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.amaze.filemanager.fragments;
+package com.amaze.filemanager.ui.fragments;
 
 import com.amaze.filemanager.R;
 import com.amaze.filemanager.activities.MainActivity;
