+++ app/src/main/java/com/amaze/filemanager/ui/fragments/preference_fragments/FoldersPref.java
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.amaze.filemanager.fragments.preference_fragments;
+package com.amaze.filemanager.ui.fragments.preference_fragments;
 
 import java.util.HashMap;
 import java.util.Map;
